==> notebooks/02-data_generation.ipynb <==
import pandas as pd
import re

# Dataset info

df = pd.read_csv('../data/raw/NLP training data 0816.csv')
df.drop_duplicates(inplace=True)
df.reset_index(inplace=True, )
df.head()

# Generation

## which plans

df.shape, df.question_description.unique().shape

mask = df.question_description.str.contains(']', regex=False)
df[mask]

mask = df.question_description.str.contains('Which plan', regex=False)
df[mask].shape, df[mask].sort_values(by='question_description').question_description.unique().shape

df[mask].question_description.unique()

processed_df = df.copy()

def which_plan(line):
    
    
    if 'and what\'s the cost' in line['question_description']:
        line['question_description'] = re.sub(r"(?P<q>Which plan\(s\) cover) (?P<entity>.[^?]+)(?P<e> and what's the cost\?)", r'\g<q> [\g<entity>]\g<e>', line['question_description'])
        line['question_type'] = 0
    
    elif 'Which plan' in line['question_description']:
        if 'coverage?' not in line['question_description']:
            line['question_description'] = re.sub(r'(?P<q>cover|offer) (?P<entity>.[^?]+)', r'\g<q> [\g<entity>]', line['question_description'])  
        else:
            line['question_description'] = re.sub(r'(?P<q>cover|offer|have) (?P<entity>.[^?]+)(?P<e> coverage\?)', r'\g<q> [\g<entity>]\g<e>', line['question_description'])  
        line['question_type'] = 1
    
    elif 'How much does ' in line['question_description']:
        line['question_description'] = re.sub(r'(?P<q>(How much does)) (?P<entity>.[^?]+)(?P<e> cost to me\?)', r'\g<q> [\g<entity>]\g<e>', line['question_description'])
        line['question_type'] = 2
    
    elif 'What does ' in line['question_description']:
        line['question_description'] = re.sub(r'(?P<q>(What does)) (?P<entity>.[^?]+)(?P<e> mean\?)', r'\g<q> [\g<entity>]\g<e>', line['question_description'])
        line['question_type'] = 3
    
    elif 'What is the ' in line['question_description']:
        line['question_description'] = re.sub(r'(?P<q>(What is the)) (?P<entity>.[^?]+)', r'\g<q> [\g<entity>]', line['question_description'])
        line['question_type'] = 3
        
    elif 'I\'d like to talk to ' in line['question_description']:
        line['question_description'] = re.sub(r'(?P<q>(I\'d like to talk to)) (?P<entity>.[^?]+)', r'\g<q> [\g<entity>]', line['question_description'])
        line['question_type'] = 4
    
    return line
processed_df = processed_df.apply(which_plan, axis='columns')

mask = processed_df.question_description.str.contains('Which plan', regex=False)
processed_df[mask].question_description.unique()

mask = processed_df.question_description.str.contains('How much does ', regex=False)
processed_df[mask].question_description.unique()

mask = ~processed_df.question_description.str.contains(']', regex=False)
processed_df[mask].question_description.unique()

# Merge keywords and phrases

processed_df.head()

merged_df = processed_df.copy()
merged_df.drop(['campaign_id', 'client_id', 'index', 'question_category'], axis='columns', inplace=True)

def uncapitalize(s):
    return s[:1].lower() + s[1:] if s else ''
    
def merge_data(line):
    if isinstance(line['question_keyword'], str):
        line['train_phase'] = re.sub(r'\[[^?]+\]', uncapitalize(line['question_keyword']), line['question_description'])
    else:
        line['train_phase'] = line['question_description']
    return line
merged_df = merged_df.apply(merge_data, axis='columns')
merged_df.head()

merged_df.info()

merged_df[merged_df.train_phase.str.contains('\[')]

merged_df.question_type.value_counts()

merged_df.to_csv('../data/interim/processed_dataset.csv', index=False)

# Live agent

df = pd.read_csv('../data/external/rasa_ask_human.csv')
df.head()

df.drop_duplicates(inplace=True)
df.to_csv('../data/external/rasa_ask_human_unique.csv', index=False)
