==> notebooks/03-extract_db.ipynb <==
from getpass import getpass
from mysql.connector import connect, Error

import pandas as pd

query = "SELECT * FROM Standard_Question_Info"

try:
    with connect(database='plan_benefit_data', 
                 user='andrey', 
                 password='lFog0FCMvzktuox', 
                 host='50.116.37.183', 
                 port='3306') as connection:
        print(connection)
        
        df = pd.read_sql(query, connection)
except Error as e:
    print(e)

df

df.info()

df.to_csv('../data/raw/db_data.csv', index=False)
