==> notebooks/04-merge_dataset.ipynb <==
import pandas as pd
from random import sample, choice

# Load data

df = pd.read_csv('../data/interim/processed_dataset.csv')
df.drop_duplicates(inplace=True)
df.reset_index(inplace=True, )
df.drop(['index', 'question_subcategory', 'train_phase'], axis=1, inplace=True)
df.head()

# Load phrases

phrase_df = pd.DataFrame({'question':[], 'question_type':[]})
phrase_df['question_type'] = phrase_df['question_type'].astype(int)
for i in range(0, 5):
    temp_df = pd.read_excel('../data/external/paraphrased_questions.xlsx', sheet_name=f"question_type_{i}")
    temp_df['question_type'] = int(i)
    phrase_df = pd.concat(objs=[phrase_df, temp_df])
phrase_df['question'] = phrase_df['question'].str.replace('…', '...', regex=False)
phrase_df['insert'] = phrase_df['question'].str.contains('...', regex=False) 
print(phrase_df.info())
phrase_df.head()

phrase_df.question_type.value_counts(), phrase_df['insert'].value_counts(dropna=False)


# Merge data

def uncapitalize(s):
    return s[:1].lower() + s[1:] if s else ''

def keyword_aug(s):
    words = s.split()
    if choice([0,1]) and len(words)>5:
        range_top = choice(list(range(6, 15)))
        s = " ".join(words[:range_top])
    if choice([0,1]):
        s = s.lower()
    else:
        s = uncapitalize(s) if not s[1].isupper() else s
    return s

merged_dataset = pd.DataFrame({'data':[],'label':[]})

# dataset_list = []
# for idx, row in df.iteritems():
#     data_dict = {}
#     question_type = row['question_type']
#     mask = phrase_df.question_type == question_type
#     phrases = phrase_df[mask].

dataset_list = []
for type_idx in range(5):
    
    
    mask = (df.question_type == type_idx) & (~df.question_keyword.isnull())
    target_questions = df[mask].copy()
    if type_idx>0:
        mask = phrase_df.question_type == type_idx
        target_phrases = phrase_df[mask].question.values
        insert_flags = phrase_df[mask].question_type.values
    
        for _, row in target_questions.iterrows():
            
            
            k = row['question_description']
            k = k.replace("[", "").replace("]", "").replace("(", "").replace(")", "")
            l = row['question_id']
            data_dict = {}
            data_dict['question'] = k
            data_dict['label'] = l
            data_dict['question_type'] = type_idx
            dataset_list.append(data_dict)
            
            k = row['question_keyword']
            l = row['question_id']
            data_dict = {}
            data_dict['question'] = k
            data_dict['label'] = l
            data_dict['question_type'] = type_idx
            dataset_list.append(data_dict)
            
            for phrase, insert_flag in zip(target_phrases, insert_flags):
                q = phrase.replace('...', keyword_aug(k))
                
                data_dict = {}
                data_dict['question'] = q
                data_dict['label'] = l
                data_dict['question_type'] = type_idx
                dataset_list.append(data_dict)
                
    elif type_idx==0:
        mask = phrase_df.question_type == 1
        target_phrases_left = phrase_df[mask].question.values
        insert_flags_left = phrase_df[mask].question_type.values
        
        mask = phrase_df.question_type == 2
        target_phrases_right = phrase_df[mask].question.str.lower().values
        insert_flags_right = phrase_df[mask].question_type.values
    
        for _, row in target_questions.iterrows():
            
            
            k = row['question_description']
            k = k.replace("[", "").replace("]", "").replace("(", "").replace(")", "")
            l = row['question_id']
            data_dict = {}
            data_dict['question'] = k
            data_dict['label'] = l
            data_dict['question_type'] = type_idx
            dataset_list.append(data_dict)
            
            k = row['question_keyword']
            l = row['question_id']
            data_dict = {}
            data_dict['question'] = k
            data_dict['label'] = l
            data_dict['question_type'] = type_idx
            dataset_list.append(data_dict)
            
            for phrase_left in target_phrases_left:
                phrase_right = choice(target_phrases_right)
                q = phrase_left.replace('...', keyword_aug(k)).replace('?','') + ' and ' + phrase_right.replace('...', choice(['it', 'this', 'the', '']))
                
                data_dict = {}
                data_dict['question'] = q
                data_dict['label'] = l
                data_dict['question_type'] = type_idx
                dataset_list.append(data_dict)
            

dataset_df = pd.DataFrame(dataset_list)
dataset_df.drop_duplicates(inplace=True)
print(dataset_df.info())
dataset_df.head()

dataset_df.to_csv('../data/processed/dataset.csv', index=False)

dataset_df.label.value_counts()

dataset_df.info()
