==> notebooks/05-dataloader.ipynb <==
import pandas as pd
from sklearn.preprocessing import LabelEncoder
from src.data.dataloader import OHLDataModule

import pytorch_lightning as pl
from lightning_transformers.task.nlp.text_classification import (
    TextClassificationDataModule,
    TextClassificationTransformer,
)
from tqdm.notebook import tqdm

dataset_df = pd.read_csv('../data/processed/dataset.csv')
dataset_df.head()

pd.DataFrame([dataset_df.label.value_counts(), dataset_df.label.value_counts(normalize=True)]).T.head(20)


mask = (dataset_df.label.value_counts() < 21)

dataset_df.label.value_counts()[mask].sum(),dataset_df.label.value_counts(normalize=True)[mask].sum()*100, mask.sum()

l_encoder = LabelEncoder()
l_encoder = l_encoder.fit(dataset_df['label'].values)

l_encoder.transform(dataset_df['label'])

from transformers import AutoTokenizer

tokenizer = AutoTokenizer.from_pretrained("distilbert-base-uncased")

def preprocess_function(text):
    return tokenizer.encode_plus(text, padding='max_length', truncation=True, max_length=256)
len(preprocess_function('i wan%l to ta1k to someone at company')['input_ids'])

datamodule = OHLDataModule(data_path='../data/processed/dataset.csv',
                           max_length=512,
                           train_bs=32,
                           valid_bs=32,
                           test_size=0.3,
                           model_name='distilbert-base-uncased',
                           aug=True)
datamodule.setup()

datamodule.train_ds.texts[0], datamodule.train_ds.__getitem__(0)['labels'].shape

dataloader = datamodule.train_dataloader()

next(dataloader._get_iterator())['input_ids'].shape

model = TextClassificationTransformer(pretrained_model_name_or_path="bert-base-uncased", num_labels=datamodule.num_classes)
# model.num_classes = 
trainer = pl.Trainer(accelerator="auto", devices="auto", max_epochs=1)

trainer.fit(model, datamodule)

pd.Series({'id1':1, "id2":2})

lens = []
for idx, row in tqdm(dataset_df.iterrows(), total=dataset_df.shape[0]):
    lens.append(len(tokenizer(row['question'])['input_ids']))
