==> notebooks/06-aug.ipynb <==
import nlpaug.flow as naf
import nlpaug.augmenter.char as nac
import nlpaug.augmenter.word as naw
from src.data.augmentation import NLPTransform, OneOfTransfroms, ComposeAug

transform = [
    OneOfTransfroms([
        NLPTransform(nac.KeyboardAug(aug_char_max=1, 
                                     aug_word_max=3), 
                     p=1),
        NLPTransform(naw.SpellingAug(aug_max=3), 
                     p=1)],
        p=0.5),]
aug = ComposeAug(transform)
aug('i want to talk to someone at company')

aug

aug.to_dict()

aug = ComposeAug([])
aug
