==> notebooks/07-model.ipynb <==
from transformers import AutoTokenizer,AutoModelForSequenceClassification
import torch

tokenizer = AutoTokenizer.from_pretrained("distilbert-base-uncased")

model = AutoModelForSequenceClassification.from_pretrained("distilbert-base-cased", num_labels=249)

tok = tokenizer('very short sentence')
tok['input_ids'] = torch.IntTensor(tok['input_ids']).unsqueeze(0)
tok['attention_mask'] = torch.FloatTensor(tok['attention_mask']).unsqueeze(0)
tok['labels'] = torch.LongTensor([1]).unsqueeze(0)
tok

model.train()
output = model(**tok)

output

output.logits

torch.argmax(output.logits, dim=-1)

torch.softmax(output.logits, dim=1)

output.logits.type(torch.int).argmax(dim=-1)

output.logits.type(torch.int).argmax(dim=-1).unsqueeze(0).shape

target = torch.tensor([0, 1, 2])
preds = torch.tensor([[0.1, 0.9, 0], [0.3, 0.1, 0.6], [0.2, 0.5, 0.3]])
target.shape, preds.shape

preds.unsqueeze(-1).squeeze(-1).shape
