==> notebooks/08-predict.ipynb <==
from src.models.predict_model import OHLProductionModel
import torch

MODEL_ID = "../models/chkpts/07-09-2022_12:38:02/epoch=4-val_f1=0.95-val_accuracy=0.99-val_loss=0.0226.ckpt"
DEVICE = "cpu"

def create_model(device: str, model_id: str):

    model = OHLProductionModel(n_classes=296, device=device)
    state_dict = torch.load(model_id, map_location=torch.device(model.device))[
        "state_dict"
    ]
    model.load_state_dict(state_dict)
    model.eval().to(model.device)
    model.warmup()
    return model
model = create_model(device=DEVICE, model_id=MODEL_ID)


class_names = []
with open("../data/interim/class_names.txt", "r") as f:
    for line in f:
        class_names.append(line.strip())


input = model.preprocess_input(['blood count', 'i want to speak to a manager'],)
pred = model(input).cpu().detach().numpy()


idx = pred.argsort(axis=1) # descending sort
pred_class_name = class_names[idx[0][::-1][0]] # the class with the highest probability for the first text
print(pred_class_name)
pred_class_name = class_names[idx[1][::-1][0]] # for the second text
print(pred_class_name)
